refactor(frontend): move settings-service into api folder and update imports
